xrRender: Blenders: Model_EbB separated
