+++ src/Layers/xrRender/blenders/Blender_Model_EbB.h
@@ -1,9 +1,3 @@
-// BlenderDefault.h: interface for the CBlenderDefault class.
-//
-//////////////////////////////////////////////////////////////////////
-
-#if !defined(AFX_BLENDERDEFAULT_H__C12F64EE_43E7_4483_9AC3_29272E0401E7__INCLUDED_21)
-#define AFX_BLENDERDEFAULT_H__C12F64EE_43E7_4483_9AC3_29272E0401E7__INCLUDED_21
 #pragma once
 
 class CBlender_Model_EbB : public IBlender
@@ -14,15 +8,12 @@ public:
     xrP_BOOL oBlend;
 
 public:
-    virtual LPCSTR getComment() { return "MODEL: env^base"; }
-    virtual BOOL canBeLMAPped() { return FALSE; }
-    virtual void Save(IWriter& fs);
-    virtual void Load(IReader& fs, u16 version);
+    LPCSTR getComment() override;
+    void Save(IWriter& fs) override;
+    void Load(IReader& fs, u16 version) override;
 
-    virtual void Compile(CBlender_Compile& C);
+    void Compile(CBlender_Compile& C) override;
 
     CBlender_Model_EbB();
-    virtual ~CBlender_Model_EbB();
+    ~CBlender_Model_EbB() override = default;
 };
-
-#endif // !defined(AFX_BLENDERDEFAULT_H__C12F64EE_43E7_4483_9AC3_29272E0401E7__INCLUDED_2)
