+++ src/Layers/xrRenderPC_GL/xrRender_GL.vcxproj.filters
@@ -258,9 +258,6 @@
     <ClCompile Include="..\xrRender\blenders\Blender_Lm(EbB).cpp">
       <Filter>Shading templates\R1</Filter>
     </ClCompile>
-    <ClCompile Include="..\xrRender\blenders\Blender_Model_EbB.cpp">
-      <Filter>Shading templates\R1</Filter>
-    </ClCompile>
     <ClCompile Include="..\xrRender\blenders\Blender_Screen_SET.cpp">
       <Filter>Shading templates\R1</Filter>
     </ClCompile>
@@ -681,6 +678,9 @@
     <ClCompile Include="..\xrRender\blenders\Blender_detail_still_deferred.cpp">
       <Filter>Shading templates</Filter>
     </ClCompile>
+    <ClCompile Include="..\xrRender\blenders\Blender_Model_EbB_deferred.cpp">
+      <Filter>Shading templates</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="stdafx.h">
@@ -786,9 +786,6 @@
       <Filter>Shading templates\R1</Filter>
     </ClInclude>
     <ClInclude Include="..\xrRender\blenders\Blender_Lm(EbB).h">
-      <Filter>Shading templates\R1</Filter>
-    </ClInclude>
-    <ClInclude Include="..\xrRender\blenders\Blender_Model_EbB.h">
       <Filter>Shading templates\R1</Filter>
     </ClInclude>
     <ClInclude Include="..\xrRender\blenders\Blender_Screen_SET.h">
@@ -1125,6 +1122,9 @@
       <Filter>Core</Filter>
     </ClInclude>
     <ClInclude Include="..\xrRender\blenders\Blender_detail_still.h">
+      <Filter>Shading templates</Filter>
+    </ClInclude>
+    <ClInclude Include="..\xrRender\blenders\Blender_Model_EbB.h">
       <Filter>Shading templates</Filter>
     </ClInclude>
   </ItemGroup>
